Update #show-grid text when squares are created
When a new grid is created with user entering a new number, text
content of #show-grid does not change along with it.

For example, When the grid is hidden, the text content of the button is
'Show Grid'. When a new grid is created, it shows the grid again but
does not change the text content of the button.

Then I added a text content property inside the create squares function
to change the button back to default.

I want to show the grid when the new grid is created because that way
user knows the amount of squares. He/She can then toggle the grid on or
off.